refactor: playground page
also disable compata data for secondary formatters for now
